Fix problem resulting from non-conformity of API response to spec
The field 'byline' inside the 'docs' array is specified to be a
JSON object ({}). However, in some 'docs' items it is an empty
JSON array ([]). This results in the Retrofit call failing.

The problem has been fixed by just removing the 'byline' field
from the 'doc' model.

The problem should be addressed later again to solve it in a
cleaner way.